+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Fix filtering
 - Refresh on window focus if autoRefresh is enabled
+- Add open event in browser
 
 # v0.0.5 - 2019-10-25
 - Handle all day events better